+++ geozero/src/gdal/mod.rs
@@ -1,7 +1,9 @@
 //! GDAL conversions.
+mod gdal_error;
 pub(crate) mod gdal_reader;
 pub(crate) mod gdal_writer;
 
+pub use gdal_error::GdalError;
 pub use gdal_reader::*;
 pub use gdal_writer::*;
 
@@ -24,8 +26,7 @@ pub(crate) mod conversion {
             self.to_gdal_ndim(CoordDimensions::default())
         }
         fn to_gdal_ndim(&self, dims: CoordDimensions) -> Result<Geometry> {
-            let mut gdal = GdalWriter::new();
-            gdal.dims = dims;
+            let mut gdal = GdalWriter::with_dims(dims);
             self.process_geom(&mut gdal)?;
             Ok(gdal.geom)
         }