+++ geozero/src/postgis/postgis_sqlx.rs
@@ -33,6 +33,28 @@ impl<'de, T: FromWkb + Sized> Decode<'de, Postgres> for wkb::Decode<T> {
     }
 }
 
+impl sqlx::Type<Postgres> for wkb::Ewkb {
+    fn type_info() -> PgTypeInfo {
+        PgTypeInfo::with_name("geometry")
+    }
+}
+
+impl PgHasArrayType for wkb::Ewkb {
+    fn array_type_info() -> PgTypeInfo {
+        PgTypeInfo::with_name("_geometry")
+    }
+}
+
+impl<'de> Decode<'de, Postgres> for wkb::Ewkb {
+    fn decode(value: PgValueRef<'de>) -> Result<Self, BoxDynError> {
+        if value.is_null() {
+            return Ok(wkb::Ewkb(Vec::new()));
+        }
+        let blob = <&[u8] as Decode<Postgres>>::decode(value)?;
+        Ok(wkb::Ewkb(blob.to_vec()))
+    }
+}
+
 impl<T: GeozeroGeometry + Sized> sqlx::Type<Postgres> for wkb::Encode<T> {
     fn type_info() -> PgTypeInfo {
         PgTypeInfo::with_name("geometry")
@@ -48,9 +70,13 @@ impl<T: GeozeroGeometry + Sized> PgHasArrayType for wkb::Encode<T> {
 impl<T: GeozeroGeometry + Sized> Encode<'_, Postgres> for wkb::Encode<T> {
     fn encode_by_ref(&self, buf: &mut PgArgumentBuffer) -> IsNull {
         let mut wkb_out: Vec<u8> = Vec::new();
-        let mut writer = wkb::WkbWriter::new(&mut wkb_out, wkb::WkbDialect::Ewkb);
-        writer.dims = self.0.dims();
-        writer.srid = self.0.srid();
+        let mut writer = wkb::WkbWriter::with_opts(
+            &mut wkb_out,
+            wkb::WkbDialect::Ewkb,
+            self.0.dims(),
+            self.0.srid(),
+            Vec::new(),
+        );
         self.0
             .process_geom(&mut writer)
             .expect("Failed to encode Geometry");
@@ -121,10 +147,13 @@ macro_rules! impl_sqlx_postgis_encode {
             ) -> sqlx::encode::IsNull {
                 use $crate::GeozeroGeometry;
                 let mut wkb_out: Vec<u8> = Vec::new();
-                let mut writer =
-                    $crate::wkb::WkbWriter::new(&mut wkb_out, $crate::wkb::WkbDialect::Ewkb);
-                writer.dims = self.dims();
-                writer.srid = self.srid();
+                let mut writer = $crate::wkb::WkbWriter::with_opts(
+                    &mut wkb_out,
+                    $crate::wkb::WkbDialect::Ewkb,
+                    self.dims(),
+                    self.srid(),
+                    Vec::new(),
+                );
                 self.process_geom(&mut writer)
                     .expect("Failed to encode Geometry");
                 buf.extend(&wkb_out); // Is there a way to write directly into PgArgumentBuffer?