+++ geozero/tests/svg.rs
@@ -9,8 +9,8 @@ use std::io::Write;
 
 #[test]
 fn json_to_svg() -> Result<()> {
-    let mut f = File::open("tests/data/places.json")?;
-    let svg = GeoJsonReader(&mut f).to_svg().unwrap();
+    let f = File::open("tests/data/places.json")?;
+    let svg = GeoJsonReader(f).to_svg().unwrap();
     println!("{svg}");
     assert_eq!(
         &svg[svg.len() - 100..],
@@ -32,12 +32,7 @@ fn invert_y(header: &Header) -> bool {
     false
 }
 
-fn svg_writer<'a, W: Write>(
-    header: &Header,
-    width: u32,
-    height: u32,
-    out: &'a mut W,
-) -> SvgWriter<'a, W> {
+fn svg_writer<W: Write>(header: &Header, width: u32, height: u32, out: W) -> SvgWriter<W> {
     let mut svg = SvgWriter::new(out, invert_y(header));
     if let Some(envelope) = header.envelope() {
         svg.set_dimensions(