+++ geozero/CHANGELOG.md
@@ -1,4 +1,35 @@
+## 0.11.0 (2023-08-28)
+
+* Add support for raw WKB DB queries
+* Fix MVT large geometry processing (#151)
+* Breaking: Implement screen coord translation for MVT writer (#150)
+* Add Ewkt dialect (#155)
+* Add support for Spatialite and MySQL WKB dialects (#153)
+* Breaking: remove `set_dims` from CSVWriter. Instead use `CsvWriter::with_dims` constructor
+* Format writers can now own their input, previously only mutable borrows were allowed
+  * `CsvWriter`
+  * `GeoJsonWriter`
+  * `WkbWriter`
+  * `WktWriter`
+  * `SVGWriter`
+* Remove lifetime param from WktReader and GpxReader
+* Breaking: Replace pub fields of writers with constructors (#163, #158)
+* Add GDAL read support for more types (#165)
+* Fix invalid geometry collection output for CSV's (#167)
+* Fix empty geometry handling for WKT, WKB, GeoJSON and CSV
+* Add reader for line delimited geojson (.geojsonl) (#168)
+* Update gdal to 0.16
+
+## 0.10.0 (2023-07-07)
+
+* Remove lifetime from `GeoJsonReader`
+* Refactor MVT and GDAL errors (#148)
+* Update sqlx to 0.7
+* Update gpx to 0.9
+* Update gdal to 0.15
+
 ## 0.9.9 (2023-04-25)
+
 * For `with-mvt`, update `dup-indexer` crate to 0.3
 
 ## 0.9.8 (2023-04-09)
@@ -11,6 +42,7 @@
 * Matched breaking changes in arrow2 0.17 and lyon 1.0.1
 
 ### Internal
+
 * CI now covers all testing including postgres testing
 * MVT tests and benchmarks framework